Backend/requirements.txt
langgraph
langchain-groq
langchain-huggingface
langchain
langchain-community
pydantic
requests
python-dotenv
fastapi
uvicorn
httpx
psycopg2-binary
sentence-transformers
faiss-cpu
pypdf
python-socketio[client]
elevenlabs
gtts
pydub
python-multipart
sqlalchemy
resend